// containers/Plans/types/plan.types.ts
import { BaseOrder, OrderDetail } from '@app/containers/Orders/types/order.types'
import { Customer } from '@app/containers/Customers/types/customer.types'
import { BaseStage, Stage } from '@app/containers/Stages/types/stage.types'
import { BaseMachine } from '@app/containers/Machines/types/machine.types'
import { Tool } from '@app/containers/Tools/types/tool.types'
import { BaseProduct } from '@app/containers/Products/types/product.types'

export interface Plan {
  id: string
  code: string
  orders: BaseOrder[]
  customers: Customer[]
  status: PlAN_STATUS
  type: string
  detail: PlanDetail[]
}

export interface PlanDetail {
  id: string
  stage: BaseStage
  schedule: Schedule[]
}

export interface Schedule {
  id: string | null
  startDate: string // ISO 8601 date string
  endDate: string // ISO 8601 date string
  projectStartDate: string // ISO 8601 date string
  projectEndDate: string // ISO 8601 date string
  order: BaseOrder
  customer: Customer
  orderDetail: OrderDetail
  employee: number | null
  machines: BaseMachine[]
  tools: ScheduleTool[]
  outputProduct: BaseProduct
  weightQuota: number
}

export interface ScheduleTool extends Omit<Tool, 'stage'> {
  stage: BaseStage | null
}

export interface GetPlanByOrderIdsParams {
  list: string[]
}

export interface GetPlanDetailQuery {
  id: string
}

export interface UpdatePlanDetailPayload extends PlanDetail {}
export interface CreatePlanDetailPayload extends Omit<PlanDetail, 'id'> {}
export interface CreatePlanDetailParams {
  planId: string
}

export interface UpdatePlanDetailParams extends CreatePlanDetailParams {}

export interface DeletePlanDetailParams {
  id: string[]
}

export interface UpdatePlanStatusQuery {
  planId: string
  planStatus: PlAN_STATUS
}

export enum PlAN_STATUS {
  USING = 'using',
  PLANNING = 'planning'
}

export interface AutoGenerateDailyReportQueries {
  date: string
  employeeId: string
}